+++ lib/modules/platform/gitea/index.spec.ts
@@ -16,7 +16,9 @@ import * as memCache from '../../../util/cache/memory/index.ts';
 import * as repoCache from '../../../util/cache/repository/index.ts';
 import { toBase64 } from '../../../util/string.ts';
 import { parseUrl } from '../../../util/url.ts';
-import * as helper from '../gitea-common/helper.ts';
+import type { EnsureIssueConfig, RepoParams } from '../index.ts';
+import * as helper from './gitea-helper.ts';
+import * as gitea from './index.ts';
 import type {
   Comment,
   CommitStatus,
@@ -27,9 +29,7 @@ import type {
   Repo,
   RepoPermission,
   User,
-} from '../gitea-common/schema.ts';
-import type { EnsureIssueConfig, RepoParams } from '../index.ts';
-import * as gitea from './index.ts';
+} from './schema.ts';
 
 /**
  * latest tested gitea version.
