+++ lib/workers/repository/update/pr/changelog/gitea/index.ts
@@ -3,7 +3,7 @@ import { logger } from '../../../../../../logger/index.ts';
 import { Releases } from '../../../../../../modules/datasource/gitea-releases/schema.ts';
 import {
   ContentsListResponse,
-  ContentsResponse,
+  RepoContents,
 } from '../../../../../../modules/platform/gitea/schema.ts';
 import { GiteaHttp } from '../../../../../../util/http/gitea.ts';
 import { fromBase64 } from '../../../../../../util/string.ts';
@@ -37,7 +37,7 @@ export async function getReleaseNotesMd(
     )
   ).body;
   const allFiles = tree.filter((f) => f.type === 'file');
-  let files: ContentsResponse[] = [];
+  let files: RepoContents[] = [];
   if (!files.length) {
     files = allFiles.filter((f) => changelogFilenameRegex.test(f.name));
   }
@@ -58,10 +58,10 @@ export async function getReleaseNotesMd(
 
   const fileRes = await http.getJson(
     `${apiPrefix}/${changelogFile}`,
-    ContentsResponse,
+    RepoContents,
   );
   // istanbul ignore if: should never happen
-  if (!fileRes.body.content) {
+  if (fileRes.body.type !== 'file' || !fileRes.body.content) {
     logger.debug(`Missing content for changelog file, using ${changelogFile}`);
     return null;
   }
